Split fluent controller into two reconcilers
