+++ reference-controllers/fluent-controller/cr.yaml
@@ -1,0 +1,61 @@
+apiVersion: anvil.dev/v1
+kind: FluentBit
+metadata:
+  name: fluent-bit
+  namespace: default
+spec:
+  fluentBitConfigName: fluent-bit-config
+---
+apiVersion: anvil.dev/v1
+kind: FluentBitConfig
+metadata:
+  name: fluent-bit-config
+  namespace: default
+spec:
+  config: |
+    [Service]
+        Parsers_File    parsers.conf
+    [Input]
+        Name    tail
+        Path    /var/log/containers/*.log
+        Exclude_Path    /var/log/containers/utils_default_utils-*.log
+        Refresh_Interval    10
+        Skip_Long_Lines    true
+        DB    /fluent-bit/tail/pos.db
+        DB.Sync    Normal
+        Mem_Buf_Limit    5MB
+        Parser    docker
+        Tag    kube.*
+    [Filter]
+        Name    kubernetes
+        Match    kube.*
+        Kube_URL    https://kubernetes.default.svc:443
+        Kube_CA_File    /var/run/secrets/kubernetes.io/serviceaccount/ca.crt
+        Kube_Token_File    /var/run/secrets/kubernetes.io/serviceaccount/token
+        Labels    false
+        Annotations    false
+    [Filter]
+        Name    nest
+        Match    kube.*
+        Operation    lift
+        Nested_under    kubernetes
+        Add_prefix    kubernetes_
+    [Filter]
+        Name    modify
+        Match    kube.*
+        Remove    stream
+        Remove    kubernetes_pod_id
+        Remove    kubernetes_host
+        Remove    kubernetes_container_hash
+    [Filter]
+        Name    nest
+        Match    kube.*
+        Operation    nest
+        Wildcard    kubernetes_*
+        Nest_under    kubernetes
+        Remove_prefix    kubernetes_
+    [Output]
+        Name    kafka
+        Match_Regex    (?:kube|service)\.(.*)
+        Brokers    my-cluster-kafka-brokers.kafka.svc:9092
+        Topics    fluent-log
