+++ reference-controllers/fluent-controller/crd.yaml
@@ -1,0 +1,69 @@
+apiVersion: apiextensions.k8s.io/v1
+kind: CustomResourceDefinition
+metadata:
+  name: fluentbits.anvil.dev
+spec:
+  group: anvil.dev
+  names:
+    categories: []
+    kind: FluentBit
+    plural: fluentbits
+    shortNames: []
+    singular: fluentbit
+  scope: Namespaced
+  versions:
+    - additionalPrinterColumns: []
+      name: v1
+      schema:
+        openAPIV3Schema:
+          description: "Auto-generated derived type for FluentBitSpec via `CustomResource`"
+          properties:
+            spec:
+              properties:
+                fluentBitConfigName:
+                  type: string
+              required:
+                - fluentBitConfigName
+              type: object
+          required:
+            - spec
+          title: FluentBit
+          type: object
+      served: true
+      storage: true
+      subresources: {}
+---
+apiVersion: apiextensions.k8s.io/v1
+kind: CustomResourceDefinition
+metadata:
+  name: fluentbitconfigs.anvil.dev
+spec:
+  group: anvil.dev
+  names:
+    categories: []
+    kind: FluentBitConfig
+    plural: fluentbitconfigs
+    shortNames: []
+    singular: fluentbitconfig
+  scope: Namespaced
+  versions:
+    - additionalPrinterColumns: []
+      name: v1
+      schema:
+        openAPIV3Schema:
+          description: "Auto-generated derived type for FluentBitConfigSpec via `CustomResource`"
+          properties:
+            spec:
+              properties:
+                config:
+                  type: string
+              required:
+                - config
+              type: object
+          required:
+            - spec
+          title: FluentBitConfig
+          type: object
+      served: true
+      storage: true
+      subresources: {}
